chore: restructure Windows-compatible codes
