+++ src/client/account.rs
@@ -28,49 +28,3 @@ impl VecSerialize<Account> for Vec<Account> {
         serde_json::to_string(&self).expect("Failed to serialize Vec<Descriptor>")
     }
 }
-
-pub mod theme {
-    #[cfg(target_os = "windows")]
-    use windows::{
-        core::*,
-        Win32::System::Registry::*,
-    };
-
-    /// Use windows crate to detect Windows system dark mode as gtk settings
-    /// does not respect it
-    #[cfg(target_os = "windows")]
-    pub fn is_system_dark_mode_enabled() -> bool {
-        #[cfg(windows)]
-        unsafe {
-            let subkey = w!("Software\\Microsoft\\Windows\\CurrentVersion\\Themes\\Personalize");
-            let mut key_handle = HKEY::default();
-
-            let result = RegOpenKeyExW(HKEY_CURRENT_USER, subkey, 0, KEY_READ, &mut key_handle);
-
-            if result.is_err() {
-                return false;
-            }
-
-            let mut data: u32 = 0;
-            let mut data_size: u32 = std::mem::size_of::<u32>() as u32;
-            let value_name = w!("AppsUsesLightTheme");
-
-            let result = RegQueryValueExW(
-                key_handle,
-                value_name,
-                None,
-                None,
-                Some(&mut data as *mut u32 as *mut u8),
-                Some(&mut data_size),
-            );
-
-            let _ = RegCloseKey(key_handle);
-
-            if result.is_err() {
-                return false;
-            }
-
-            data == 0
-        }
-    }
-}
