+++ src/client/mod.rs
@@ -4,6 +4,8 @@ pub mod error;
 pub mod proxy;
 pub mod runtime;
 pub mod structs;
+#[cfg(target_os = "windows")]
+pub mod windows_compat;
 
 pub use account::Account;
 pub use proxy::ReqClient;
